% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes2stage-package.R
\docType{package}
\name{bayes2stage-package}
\alias{bayes2stage}
\alias{bayes2stage-package}
\title{bayes2stage: Bayesian Analysis of Two-Stage Designs with Longitudinal Data}
\description{
Tools for designing and analyzing two-stage studies with longitudinal data. Supports outcome-dependent sampling (ODS) and BLUP-dependent sampling (BDS) designs where expensive covariates are measured on a subset of subjects. Bayesian model fitting uses 'Stan' via the 'instantiate' package for mixed effects models with imputation. Also provides ascertainment-corrected maximum likelihood (ACML) estimation and legacy support for 'NIMBLE'-based fitting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maxdrohde/bayes2stage}
  \item Report bugs at \url{https://github.com/maxdrohde/bayes2stage/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Rohde \email{maxdrohde@gmail.com} (\href{https://orcid.org/0000-0002-3731-3372}{ORCID})

}
\keyword{internal}
