% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bds_design.R
\name{bds_design}
\alias{bds_design}
\title{Set x_e to missing based on an BDS design}
\usage{
bds_design(
  dataset,
  fixed_effects_formula,
  sampling_type,
  cutoff_high,
  cutoff_low,
  sampling_N,
  prop_high,
  prop_middle,
  prop_low
)
}
\arguments{
\item{dataset}{Dataset to use}

\item{fixed_effects_formula}{Formula for the fixed-effects when fitting the model to estimate BLUPs}

\item{sampling_type}{Which type of sampling? "intercept" or "slope"}

\item{cutoff_high}{Which quantile to use as the cutoff for the High category}

\item{cutoff_low}{Which quantile to use as the cutoff for the Low category}

\item{sampling_N}{How many subjects should be sampled?}

\item{prop_high}{What proportion to sample from the High category?}

\item{prop_middle}{What proportion to sample from the Middle category?}

\item{prop_low}{What proportion to sample from the Low category?}
}
\value{
A dataset where the x_e values are selected based on an BDS design
}
\description{
Set x_e to missing based on an BDS design
}
