% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_cols}
\alias{check_cols}
\title{Check for required columns in a data frame}
\usage{
check_cols(data, required_cols)
}
\arguments{
\item{data}{A data frame to check}

\item{required_cols}{A character vector of required column names}
}
\value{
Invisibly returns TRUE if all columns are present; otherwise throws an error
}
\description{
Validates that a data frame contains all required columns.
}
