% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stan_data.R
\name{create_stan_data}
\alias{create_stan_data}
\title{Format the simulated data for Stan}
\usage{
create_stan_data(dataset)
}
\arguments{
\item{dataset}{Dataset to use}
}
\value{
A list suitable for input to Stan
}
\description{
Format the simulated data for Stan
}
