% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plotting.R
\name{data_plot}
\alias{data_plot}
\title{Plot simulated data}
\usage{
data_plot(dataset, subset_size = 200)
}
\arguments{
\item{dataset}{Dataset to use}

\item{subset_size}{How many subjects to use?}
}
\value{
A ggplot2 object
}
\description{
Plot simulated data
}
