% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nimble_model.R
\name{fit_nimble_model}
\alias{fit_nimble_model}
\title{Fit the model with NIMBLE}
\usage{
fit_nimble_model(
  df,
  niter = 20000,
  nburnin = 5000,
  nchains = 4,
  print_model_details = FALSE
)
}
\description{
Fit the model with NIMBLE
}
