% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_schild_2019.R
\name{generate_data_schild_2019_supp}
\alias{generate_data_schild_2019_supp}
\title{Generate correlated random intercept / slope longitudinal data
to test two-stage design methods. In the form of Schildcrout (2019)
supplement}
\usage{
generate_data_schild_2019_supp(
  N,
  Ms,
  alpha_main,
  beta_x,
  beta_z,
  beta_t,
  beta_t_x_interaction,
  beta_t_z_interaction,
  error_sd,
  x_prevalence,
  rand_intercept_sd,
  rand_slope_sd,
  rand_eff_corr,
  gamma0,
  gamma1,
  gamma2,
  gamma_sd
)
}
\arguments{
\item{N}{Integer. Total number of subjects in the cohort. Default is 2000.}

\item{Ms}{Integer vector. Possible numbers of observations per subject,
sampled uniformly. Default is c(4, 5, 6).}

\item{alpha_main}{Numeric. Fixed intercept (β₀). Default is 75.}

\item{beta_x}{Numeric. Main effect of binary exposure X (β_s). Default is -0.5.}

\item{beta_z}{Numeric. Effect of continuous confounder Z (β_c). Default is -2.}

\item{beta_t}{Numeric. Fixed time effect (β_t). Default is -1.}

\item{beta_t_x_interaction}{Numeric. Interaction between time and exposure
(β_st). Default is -0.5.}

\item{beta_t_z_interaction}{Numeric. Interaction between time and confounder.
Default is 0.}

\item{error_sd}{Numeric. Standard deviation of measurement error (σ_e).
Default is sqrt(12.25) = 3.5.}

\item{x_prevalence}{Numeric. Prevalence of binary exposure X. Must be between
0 and 1. Default is 0.25.}

\item{rand_intercept_sd}{Numeric. Standard deviation of random intercept (σ₀).
Default is sqrt(81) = 9.}

\item{rand_slope_sd}{Numeric. Standard deviation of random slope (σ₁).
Default is sqrt(1.56) = 1.25.}

\item{rand_eff_corr}{Numeric. Correlation between random intercept and slope
(ρ). Must be between -1 and 1. Default is 0.}

\item{gamma0}{Numeric. Intercept for confounder Z generation. Default is 0.25.}

\item{gamma1}{Numeric. Linear coefficient for X in Z generation. Default is 0.5.}

\item{gamma2}{Numeric. Quadratic coefficient for X in Z generation. Default is 0.}

\item{gamma_sd}{Numeric. Standard deviation of Z residual error. Default is 1.}
}
\value{
dataset
}
\description{
Generate correlated random intercept / slope longitudinal data
to test two-stage design methods. In the form of Schildcrout (2019)
supplement
}
