% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_imputation.R
\name{get_gamma_posterior_mean}
\alias{get_gamma_posterior_mean}
\title{Extract posterior means for gamma}
\usage{
get_gamma_posterior_mean(summary_df)
}
\arguments{
\item{summary_df}{Summary data frame from fit$summary()}
}
\value{
Vector of gamma posterior means
}
\description{
Extract posterior means for gamma
}
\keyword{internal}
