% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_imputation.R
\name{get_ppc_data}
\alias{get_ppc_data}
\title{Extract observed x and design matrix Z safely}
\usage{
get_ppc_data(data, imputation_covariates)
}
\arguments{
\item{data}{Full data frame}

\item{imputation_covariates}{Character vector of covariate names}
}
\value{
List with x (vector) and Z (matrix)
}
\description{
Extract observed x and design matrix Z safely
}
\keyword{internal}
