% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{mcmc_forest}
\alias{mcmc_forest}
\title{Create a forest plot of MCMC output}
\usage{
mcmc_forest(mcmc_output)
}
\arguments{
\item{mcmc_output}{MCMC output object (e.g., from fit_model)}
}
\value{
A forest plot of parameter estimates
}
\description{
A wrapper around MCMCvis::MCMCplot to create forest plots of MCMC samples.
}
