% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mcmc_trace}
\alias{mcmc_trace}
\title{Plot MCMC trace plots}
\usage{
mcmc_trace(mcmc_output, print_to_pdf = FALSE)
}
\arguments{
\item{mcmc_output}{MCMC output object (e.g., from fit_model)}

\item{print_to_pdf}{Logical; if TRUE, saves trace plots to a PDF file}
}
\value{
Trace plots of MCMC samples
}
\description{
A wrapper around MCMCvis::MCMCtrace to create trace plots of MCMC samples.
}
