% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model_summary}
\alias{model_summary}
\alias{model_summary.CmdStanMCMC}
\alias{model_summary.CmdStanPathfinder}
\alias{model_summary.CmdStanLaplace}
\alias{model_summary.CmdStanMLE}
\alias{model_summary.acml_fit}
\alias{model_summary.default}
\title{Extract summary statistics from model fit objects}
\usage{
model_summary(object, ...)

\method{model_summary}{CmdStanMCMC}(object, ...)

\method{model_summary}{CmdStanPathfinder}(object, ...)

\method{model_summary}{CmdStanLaplace}(object, ...)

\method{model_summary}{CmdStanMLE}(object, ...)

\method{model_summary}{acml_fit}(object, ...)

\method{model_summary}{default}(object, ...)
}
\arguments{
\item{object}{A model fit object}

\item{...}{Additional arguments passed to methods}
}
\value{
A data frame with summary statistics for each parameter
}
\description{
S3 generic for extracting summary statistics from various fit object types.
Supports CmdStanR fit types and ACML fits.
}
