% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ods_design.R
\name{ods_design}
\alias{ods_design}
\title{Set x to missing based on an ODS design}
\usage{
ods_design(
  data,
  sampling_type,
  cutoff_high,
  cutoff_low,
  n_sampled,
  prop_high,
  prop_middle,
  prop_low
)
}
\arguments{
\item{data}{Dataset to use}

\item{sampling_type}{Which type of sampling? "intercept" or "slope"}

\item{cutoff_high}{Which quantile to use as the cutoff for the High category}

\item{cutoff_low}{Which quantile to use as the cutoff for the Low category}

\item{n_sampled}{How many subjects should be sampled?}

\item{prop_high}{What proportion to sample from the High category?}

\item{prop_middle}{What proportion to sample from the Middle category?}

\item{prop_low}{What proportion to sample from the Low category?}
}
\value{
A dataset where the x values are selected based on an ODS design
}
\description{
Set x to missing based on an ODS design
}
