% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot simulated data}
\usage{
plot_data(data, subset_size = 200)
}
\arguments{
\item{data}{Dataset to use}

\item{subset_size}{How many subjects to use?}
}
\value{
A ggplot2 object
}
\description{
Plot simulated data
}
