% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_energy}
\alias{plot_energy}
\title{Energy diagnostic plot (NUTS-specific)}
\usage{
plot_energy(fit)
}
\arguments{
\item{fit}{CmdStanMCMC fit object}
}
\value{
A ggplot object
}
\description{
Energy diagnostic plot (NUTS-specific)
}
