% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_funnel}
\alias{plot_funnel}
\title{Quick funnel diagnostic for hierarchical parameters}
\usage{
plot_funnel(
  fit,
  sigma_param = "sigma_re[1]",
  effect_param = "re\\\\[1,",
  n_effects = 5
)
}
\arguments{
\item{fit}{CmdStanMCMC fit object}

\item{sigma_param}{Name of the sigma parameter (default: \code{"sigma_re[1]"})}

\item{effect_param}{Pattern for random effect parameters (default: \verb{"re\\\[1,"})}

\item{n_effects}{Number of random effects to plot (default: 5)}
}
\value{
A ggplot object
}
\description{
Creates a scatter plot of log(sigma) vs random effects to diagnose funnel geometry.
}
