% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_parcoord}
\alias{plot_parcoord}
\title{Parallel coordinates plot for diagnosing divergences}
\usage{
plot_parcoord(fit, parameters = NULL)
}
\arguments{
\item{fit}{CmdStanMCMC fit object}

\item{parameters}{Character vector of parameter names.
If NULL, uses parameters related to random effects.}
}
\value{
A ggplot object
}
\description{
Parallel coordinates plot for diagnosing divergences
}
