% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_pareto_k}
\alias{plot_pareto_k}
\title{Pareto k diagnostic plot for divergences}
\usage{
plot_pareto_k(fit, parameters = NULL)
}
\arguments{
\item{fit}{CmdStanMCMC fit object}

\item{parameters}{Character vector of parameter names.
If NULL, uses parameters related to random effects.}
}
\value{
A ggplot object
}
\description{
Pareto k diagnostic plot for divergences
}
