% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_imputation.R
\name{ppc_imputation}
\alias{ppc_imputation}
\title{Posterior Predictive Checks for Imputation Model}
\usage{
ppc_imputation(
  fit,
  data,
  imputation_covariates,
  distribution = c("normal", "bernoulli", "negative_binomial", "beta_binomial"),
  n_draws = 1000
)
}
\arguments{
\item{fit}{CmdStanMCMC fit object from fit_stan_model()}

\item{data}{Data frame used for fitting}

\item{imputation_covariates}{Character vector of imputation model covariate names}

\item{distribution}{Imputation distribution used}

\item{n_draws}{Number of posterior draws for simulations}
}
\value{
List with all diagnostic results and plots
}
\description{
Functions to diagnose imputation model adequacy
}
