% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_imputation.R
\name{ppc_imputation_bernoulli}
\alias{ppc_imputation_bernoulli}
\title{PPC for Bernoulli imputation model}
\usage{
ppc_imputation_bernoulli(fit, data, imputation_covariates, n_draws = 1000)
}
\arguments{
\item{fit}{CmdStanMCMC fit object}

\item{data}{Data frame}

\item{imputation_covariates}{Character vector of imputation model covariates}

\item{n_draws}{Number of posterior draws}
}
\value{
List with calibration plot, ROC metrics
}
\description{
PPC for Bernoulli imputation model
}
