% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_imputation.R
\name{ppc_imputation_beta_binomial}
\alias{ppc_imputation_beta_binomial}
\title{PPC for beta-binomial imputation model}
\usage{
ppc_imputation_beta_binomial(fit, data, imputation_covariates, n_draws = 500)
}
\arguments{
\item{fit}{CmdStanMCMC fit object}

\item{data}{Data frame}

\item{imputation_covariates}{Character vector of imputation model covariates}

\item{n_draws}{Number of posterior draws}
}
\value{
List with diagnostics
}
\description{
PPC for beta-binomial imputation model
}
