% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_imputation.R
\name{ppc_imputation_count}
\alias{ppc_imputation_count}
\title{PPC for count imputation model (negative binomial)}
\usage{
ppc_imputation_count(fit, data, imputation_covariates, n_draws = 500)
}
\arguments{
\item{fit}{CmdStanMCMC fit object}

\item{data}{Data frame}

\item{imputation_covariates}{Character vector of imputation model covariates}

\item{n_draws}{Number of posterior draws}
}
\value{
List with rootogram and dispersion check
}
\description{
PPC for count imputation model (negative binomial)
}
