% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_imputation.R
\name{ppc_imputation_density}
\alias{ppc_imputation_density}
\title{Posterior predictive density check for imputation model}
\usage{
ppc_imputation_density(fit, data, imputation_covariates, n_draws = 100)
}
\arguments{
\item{fit}{CmdStanMCMC fit object from fit_stan_model()}

\item{data}{Data frame used for fitting}

\item{imputation_covariates}{Character vector of imputation model covariates}

\item{n_draws}{Number of posterior draws to visualize}
}
\value{
ggplot object
}
\description{
Posterior predictive density check for imputation model
}
