% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_imputation.R
\name{ppc_imputation_intervals}
\alias{ppc_imputation_intervals}
\title{Check prediction interval coverage}
\usage{
ppc_imputation_intervals(
  fit,
  data,
  imputation_covariates,
  prob = 0.95,
  n_draws = 1000
)
}
\arguments{
\item{fit}{CmdStanMCMC fit object}

\item{data}{Data frame}

\item{imputation_covariates}{Character vector of imputation model covariates}

\item{prob}{Coverage probability}

\item{n_draws}{Number of posterior draws}
}
\value{
List with coverage proportion and plot
}
\description{
Check prediction interval coverage
}
