% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_imputation.R
\name{ppc_imputation_residuals}
\alias{ppc_imputation_residuals}
\title{Residual diagnostics for normal imputation model}
\usage{
ppc_imputation_residuals(fit, data, imputation_covariates)
}
\arguments{
\item{fit}{CmdStanMCMC fit object}

\item{data}{Data frame}

\item{imputation_covariates}{Character vector of imputation model covariates}
}
\value{
List with diagnostic plots
}
\description{
Residual diagnostics for normal imputation model
}
