% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.bayes2stage_diagnostics}
\alias{print.bayes2stage_diagnostics}
\title{Print method for bayes2stage diagnostics}
\usage{
\method{print}{bayes2stage_diagnostics}(x, ...)
}
\arguments{
\item{x}{bayes2stage_diagnostics object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for bayes2stage diagnostics
}
