% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_imputation.R
\name{print.ppc_imputation}
\alias{print.ppc_imputation}
\title{Print method for ppc_imputation results}
\usage{
\method{print}{ppc_imputation}(x, ...)
}
\arguments{
\item{x}{ppc_imputation results}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for ppc_imputation results
}
