% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_missing.R
\name{set_missing}
\alias{set_missing}
\title{Given selected subjects, set X to missing for those not selected}
\usage{
set_missing(data, selected_ids)
}
\arguments{
\item{data}{Input data frame}

\item{selected_ids}{IDs for the selected subjects}
}
\value{
data.frame
}
\description{
Given selected subjects, set X to missing for those not selected
}
