% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs_design.R
\name{srs_design}
\alias{srs_design}
\title{Set x_e to missing based on an SRS design}
\usage{
srs_design(dataset, proportion)
}
\arguments{
\item{dataset}{Dataset to use}

\item{proportion}{What proportion of subjects to sample?}
}
\value{
A dataset where the x_e values are selected based on an SRS design
}
\description{
Set x_e to missing based on an SRS design
}
