% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes2stage-package.R
\docType{package}
\name{bayes2stage-package}
\alias{bayes2stage}
\alias{bayes2stage-package}
\title{bayes2stage: Bayesian Analysis of Two-Stage Designs}
\description{
Provides functions for analyzing Bayesian Two-Stage Designs using NIMBLE for model fitting, and for generating simulated data from two-stage designs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maxdrohde/bayes2stage}
  \item Report bugs at \url{https://github.com/maxdrohde/bayes2stage/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Rohde \email{maxdrohde@gmail.com} (\href{https://orcid.org/0000-0002-3731-3372}{ORCID})

}
\keyword{internal}
