% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bds_design.R
\name{bds_design}
\alias{bds_design}
\title{Set x_e to missing based on an BDS design}
\usage{
bds_design(
  dataset,
  type,
  cutoff_high = 0.9,
  cutoff_low = 0.1,
  prop_high = 0.1,
  prop_middle = 0.05,
  prop_low = 0.1
)
}
\arguments{
\item{dataset}{Dataset to use}

\item{type}{Which type of sampling? "intercept" or "slope"}

\item{cutoff_high}{Which quantile to use as the cutoff for the High category}

\item{cutoff_low}{Which quantile to use as the cutoff for the Low category}

\item{prop_high}{What proportion to sample from the High category?}

\item{prop_middle}{What proportion to sample from the Middle category?}

\item{prop_low}{What proportion to sample from the Low category?}
}
\value{
A dataset where the x_e values are selected based on an BDS design
}
\description{
Set x_e to missing based on an BDS design
}
