% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_acml_ods.R
\name{fit_acml_ods}
\alias{fit_acml_ods}
\title{fit_acml_ods}
\usage{
fit_acml_ods(ods_df, cutoff_low, cutoff_high)
}
\arguments{
\item{ods_df}{ods_df}
}
\value{
params with conf ints
}
\description{
fit_acml_ods
}
