% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data_mcmc}
\alias{format_data_mcmc}
\title{Format the simulated data for Stan}
\usage{
format_data_mcmc(
  data,
  main_model_covariates = NULL,
  imputation_model_covariates = NULL,
  imputation_distribution = c("normal", "bernoulli", "beta_binomial",
    "negative_binomial")
)
}
\arguments{
\item{data}{Dataset to use}

\item{main_model_covariates}{Character vector of column names for covariates in the main model}

\item{imputation_model_covariates}{Character vector of column names for covariates in the imputation model}

\item{imputation_distribution}{Distribution for the imputation model:
"normal" for continuous x, "bernoulli" for binary x, "beta_binomial"
for bounded count data, or "negative_binomial" for unbounded count data}
}
\value{
A list suitable for input to MCMC software
}
\description{
Format the simulated data for Stan
}
