% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate correlated random intercept / slope longitudinal data
to test two-stage design methods.}
\usage{
generate_data(
  N = 2000,
  M = 5,
  alpha_main = 1,
  beta_x = 1,
  beta_z = 1,
  beta_t = 2,
  beta_t_xe_interaction = 0.3,
  error_sd = 4,
  x_dist = c("normal", "poisson", "binomial", "negative_binomial", "beta_binomial"),
  x_size = NULL,
  x_disp_param = NULL,
  rand_intercept_sd = 3,
  rand_slope_sd = 1,
  rand_eff_corr = 0,
  gamma0 = 1,
  gamma1 = 1,
  gamma2 = 0,
  gamma_sd = 2
)
}
\arguments{
\item{N}{TO DO}
}
\value{
TO DO
}
\description{
Generate correlated random intercept / slope longitudinal data
to test two-stage design methods.
}
