% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mixed_effects_data.R
\name{generate_mixed_effects_data}
\alias{generate_mixed_effects_data}
\title{Generate continuous mixed-effects data for the purpose of testing the Bayesian
Two-Stage design methods. We generate longitudinal data with random intercepts
and slopes. There are two continuous covariate, the expensive covariate
x_e and the inexpensive covariate x_z.}
\usage{
generate_mixed_effects_data(
  N = 1000,
  Ms = 5,
  alpha_main = 1,
  beta_x_e = 2,
  beta_x_z = 0.5,
  beta_t = 2,
  beta_t_xe_interaction = 1,
  error_sd = 1,
  rand_intercept_sd = 3,
  rand_slope_sd = 3,
  rand_eff_corr = 0.4,
  x_cov = 0.6,
  xe_var = 1,
  xz_var = 1
)
}
\value{
A simulated mixed-effects dataset
}
\description{
Generate continuous mixed-effects data for the purpose of testing the Bayesian
Two-Stage design methods. We generate longitudinal data with random intercepts
and slopes. There are two continuous covariate, the expensive covariate
x_e and the inexpensive covariate x_z.
}
