% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srs_design.R
\name{srs_design}
\alias{srs_design}
\title{Set x to missing based on an SRS design}
\usage{
srs_design(data, sampling_N)
}
\arguments{
\item{data}{Dataset to use}

\item{sampling_N}{How many subjects to sample?}
}
\value{
A dataset where the x values are selected based on an SRS design
}
\description{
Set x to missing based on an SRS design
}
